+++ packages/expo-media-library/CHANGELOG.md
@@ -6,6 +6,9 @@
 
 ### 🎉 New features
 
+- [Android] Add granularPermissions option to config plugin. ([#37036](https://github.com/expo/expo/pull/37036) by [@maykonmichel](https://github.com/maykonmichel))
+
+
 ### 🐛 Bug fixes
 
 - [Android] Fix `deleteAssetsAsync` not working on android 11 or above ([#33211](https://github.com/expo/expo/pull/33211) by [@Zeeshan404](https://github.com/Zeeshan404))